meus programas em python

+++ Python/006.py
@@ -1,5 +1,5 @@
 print("##### Dividindo 2 números em python #####")
-n1 = int(input("Digite um número: "))
-n2 = int(input("Digite outro número: "))
+n1 = float(input("Digite um número: "))
+n2 = float(input("Digite outro número: "))
 divisao = n1 / n2
 print("A divisão entre {} e {} é igual a {}!".format(n1,n2,divisao))
